Add type USR to the response from collect-expression and collect-vars SourceKit requests.
The USR is a useful way to uniquely identify the types. Even when the fully qualified type is provided, the USR is more useful because it's much easier to map the USR to the type in other data stores, e.g. the Indexstore.

I'm fine with adding an option to make this conditional if there are any concerns. For now, I started with just adding it directly to the response.