Emit synthesized members for protocol conformances.
For example, when Equatable conformance is synthesized also emit the decl definitions for the synthesized methods. These aren't _directly_ usable (as the synthesized methods) but those methods are used when the relevant `==` operator is used so they're useful to finding that definition.

At the moment, I'm using the location of the declaration with the conformance as the location for these definitions in the indexstore.